+++ userspace/alibc/src/link.c
@@ -2,5 +2,5 @@
 
 int link(const char *oldpath, const char *newpath)
 {
-   return asyscall(SYS_LINK, oldpath, newpath, 0, 0, 0, 0);
+   return (int)asyscall(SYS_LINK, oldpath, newpath, 0, 0, 0, 0);
 }
