+++ userspace/alibc/src/mknod.c
@@ -6,5 +6,5 @@
 
 int mknod(const char *pathname, mode_t mode, dev_t dev)
 {
-   return asyscall(SYS_MKNOD, fullpath(get_current_dir_name(), pathname), mode & ~gmask, 0, 0, 0, 0);
+   return (int)asyscall(SYS_MKNOD, fullpath(get_current_dir_name(), pathname), mode & ~gmask, 0, 0, 0, 0);
 }
