+++ userspace/alibc/src/nanosleep.c
@@ -8,6 +8,6 @@
 int nanosleep(const struct timespec *req, struct timespec* rem)
 {
    rem = NULL;
-   asyscall(SYS_USLEEP, req->tv_sec * 1000000000 + req->tv_nsec, 0, 0, 0, 0, 0);
+   (int)asyscall(SYS_USLEEP, req->tv_sec * 1000000000 + req->tv_nsec, 0, 0, 0, 0, 0);
    return 0;
 }
