+++ userspace/alibc/src/pgid.c
@@ -3,12 +3,12 @@
 
 int setpgid(pid_t pid, pid_t pgid)
 {
-   return asyscall(SYS_SETPGID, pid, pgid, 0, 0, 0, 0);
+   return (int)asyscall(SYS_SETPGID, pid, pgid, 0, 0, 0, 0);
 }
 
 pid_t getpgid(pid_t pid)
 {
-   return asyscall(SYS_GETPGID, pid, 0, 0, 0, 0, 0);
+   return (int)asyscall(SYS_GETPGID, pid, 0, 0, 0, 0, 0);
 }
 
 pid_t getpgrp(void)
