+++ userspace/alibc/src/rename.c
@@ -2,5 +2,5 @@
 
 int rename(const char *oldpath, const char *newpath)
 {
-   return asyscall(SYS_RENAME, oldpath, newpath, 0, 0, 0, 0);
+   return (int)asyscall(SYS_RENAME, oldpath, newpath, 0, 0, 0, 0);
 }
