+++ userspace/alibc/src/rlimit.c
@@ -4,7 +4,7 @@
 
 int getrlimit(int resource, struct rlimit *rlim)
 {
-   return asyscall(SYS_GETRLIMIT, resource, rlim, 0, 0, 0, 0);
+   return (int)asyscall(SYS_GETRLIMIT, resource, rlim, 0, 0, 0, 0);
 }
 
 /* NOT REALIZED */
