+++ userspace/alibc/src/sleep.c
@@ -3,5 +3,5 @@
 
 unsigned int sleep(unsigned int seconds)
 {
-   return asyscall(SYS_USLEEP, (unsigned long)(1000000 * seconds), 0, 0, 0, 0, 0);
+   return (int)asyscall(SYS_USLEEP, (unsigned long)(1000000 * seconds), 0, 0, 0, 0, 0);
 }
