+++ userspace/alibc/src/usleep.c
@@ -3,5 +3,5 @@
 
 int usleep(unsigned long usec)
 {
-   return asyscall(SYS_USLEEP, usec, 0, 0, 0, 0, 0);
+   return (int)asyscall(SYS_USLEEP, usec, 0, 0, 0, 0, 0);
 }
